executor: add the step handler
The step handler takes care of the execution of a step, using either
a built-in set of actions to be taken, or executing a user-defined
script defined in the part specification. User-defined scripts may also
call the built-in handler for a step by invoking a control utility.

This class implements the built-in handlers and a FIFO-based mechanism
and API to be used by the external control utility to communicate with
the running instance.

The control API server is a direct port of the implementation used in
Snapcraft, and should be refactored later.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>