ctl: add ctl protocol client
The control protocol client allows a user scriptlet to execute the
default handler for a step in the running application context, or set
the value of a custom variable previously passed as an argument to
LifecycleManager.

The craft_parts.ctl module can be imported by application-defined
clients, or invoked as a utility from the command line using the
supplied partsctl tool.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>