api: pass cache directory to lifecycle manager
Instead of using a path based on the application name, let the
application choose the full path for file caching. This gives the
application control over the location of the file cache, allowing
it to set the path according to its own policies and making it
easier to remove the cache if needed.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>