api: add application package name argument
Add a new `application_package_name` argument that is used to verify
if running as a snap. If not specified, it defaults to `application_name`.
This is added so that an application can run under a different snap name
such as `<application>-<username>` without affecting step environemnt
variable names, and also to allow snap names to use characters that
are not valid in `application_name`.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>