overlay: add layer state management helper
The sequencer must keep track of layer state and changes in order to
emit actions related to the overlay step (such as layer reapplications
or re-execution of steps invalidated by changes in layers). Add a
layer state management helper to be used during the planning phase
to store the layer hash corresponding to each part and compute the
expected values to ensure overlay consistency.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>