sources: fix removed file updates
Source handlers are used to pull sources from the location specified
in the part definition into the part source work directory. They may
provide a method to update specific files that have been modified
without re-pulling the entire source tree.

Changes were introduced in 5d56713 to allow removal of pulled files
that have been deleted in the part source.

The local source handler, however, is not used only for the primary
source handler purpose, it's also used to migrate contents from the
source work directory to the build work directory, regardless of the
type of the part's source. (We should investigate how to refactor this
to use the migration mechanisms directly without reusing the source
handler.)

As a consequence of the way deleted file removal was implemented and
the reuse of the local source handler for build migration, artifacts
created during the build process were removed in a source update.
Deleted file updates should cover only files removed from the original
source without affecting build artifacts.

Refactor source handlers to preserve build artifacts by removing
only entries present in the source work directory that are not
in the original source from the build work directory. In the long
term part handling should be refactored to be decoupled from the
local source handler.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>