sources: revert updating deleted files in outdated check
Commit 5d56713 is the wrong solution for the problem as it removes build
artifacts (because LocalSource is reused to migrate from `part_src_dir`
to `part_build_dir`) and any files created by the user in override-pull.

Updating deleted files can be revisited later. We should disable deleted
file synchronization during the migration to preserve build artifacts,
and be sure that the pull update handler is aware of new files created
as a part of the pull step.

For now, if the user deletes local source files they need to repull.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>