craftctl: ensure variable consistency between runs
When a project variable is set in a step, it must be available in
the next step even if execution of the first step is skipped because
it ran in a previous execution round. Write project vars to state
after step execution, read them in sequencer, and pass the expected
final state to be updated after a skipped action is processed.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>