Add groups dependency support
This patch allows to add a 'groups' statement to each part,
define one or more groups for each one, and use the groups for
dependency management.

This way, it is possible to add all the binaries and libraries
parts to a "binaries" group, and make another part deppend on
that group (just using the 'after' statement like with a part
name).

This is mainly useful in big snapcraft.yaml files with a lot
of parts, like the gnome-XX-sdk ones.

This MR replaces https://github.com/snapcore/snapcraft/pull/3823