plugins: drop cmake install prefix argument
Setting `CMAKE_INSTALL_PREFIX` by default makes the cmake plugin
compatible with the Snapcraft cmake plugin for core18 and solves
usability issues, but it also breaks compatibility with existing
core22 snaps that expect binaries to be installed in `usr/local/bin`
by default.

To keep compatibility with existing core22 snaps, the prefix path
will not be defined by default, and snap authors are encouraged
to define `-DCMAKE_INSTALL_PREFIX=` in cmake parameters to be
compatible with the core18 cmake plugin, so that libraries
built with cmake and staged by a different part can be correctly
detected without the need to define additional parameters such as
`CMAKE_INCLUDE_PATH` and `CMAKE_LIBRARY_PATH`.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>