tests: make rust integration tests use datadir
The `pytest-datadir` plugin enable per-test-module data, that is exposed
to tests via the `datadir` fixture. This proof-of-concept commit
refactors the existing integration tests for the rust plugin to use this
scheme.

The main change is that `test_rust.py` is simplified because all the
runtime rust code generation is moved to data files in the `test_rust/`
dir; this is useful because these rust files are static, and their
generation is not what's being tested.