packages: cut chisel slices when unpacking
Add the support for chisel slices but calling "chisel cut" when
unpacking stage-packages. Since we need to know which slices to cut and
we don't support fetching them individually yet (chisel limitation), use
the PullState to "persist" the list of required slices.

Note that the integrated test requires the actual 'chisel' binary so
it's only run on CI.