ci: add "chisel" dep
Chisel will only be required in some integration tests, so add it in the
gha. Note that its codebase needs a newer version of "go" than what's
available in the Ubuntu 20.04 repo (the version of ubuntu that runs the
craft-parts tests), so we

a) install a snap for a newer "go";
b) build chisel, installing it system-wide;
c) remove the "go" snap and install the regular "golang" package.