unpack chisel slices (#281)
* ci: add "chisel" dep

Chisel will only be required in some integration tests, so add it in the
gha. Note that its codebase needs a newer version of "go" than what's
available in the Ubuntu 20.04 repo (the version of ubuntu that runs the
craft-parts tests), so we

a) install a snap for a newer "go";
b) build chisel, installing it system-wide;
c) remove the "go" snap and install the regular "golang" package.

* packages: cut chisel slices when unpacking

Add the support for chisel slices but calling "chisel cut" when
unpacking stage-packages. Since we need to know which slices to cut and
we don't support fetching them individually yet (chisel limitation), use
the PullState to "persist" the list of required slices.

Note that the integrated test requires the actual 'chisel' binary so
it's only run on CI.