executor: take permissions into account for collisions (#305)
* executor: take permissions into account for collisions

Two paths are considered to be colliding if they both have permissions
definitions. The definitions are not "evaluated" to see if they possibly
match (that is, if both sets of definitions would generate the same
result) for simplicity, but this can be improved in the future.

* compare permissions to determine conflict

Do a little bit of work to determine whether two sets of Permissions are
"equivalent" or "compatible" - that is, if they both end up doing the
same thing.