plugins: add ant plugin (#332)
* plugins: add AntPlugin

The plugin requires the ``ant`` tool installed on the system. This can
be achieved by adding the appropriate declarations to ``build-packages``
or ``build-snaps``, or by having it installed or built in a different
part. In this case, the name of the part supplying ``ant`` must be
"ant-deps".

Additionally, Java projects need a dev kit (jdk) to build and a runtime
environment (jre) to run. There are multiple choices here, but frequently
adding ``default-jdk-headless`` to ``build-packages`` and
``default-jre-headless`` to ``stage-packages`` is enough.

Once built, the plugin will create the following structure in the part's
install dir (which will later be staged/primed/packaged):

- A ``bin/java`` symlink pointing to the actual ``java`` binary provided
  by the jre;
- A ``jar/`` directory containing the .jar files generated by the build.