plugins: check if plugin properties are dirty
Add plugin properties to the part state and allow the plugin to
list properties to be used to verify if the pull or build steps are
dirty.

This implementation ensures compatibility with the current plugin API
and doesn't require changes in existing plugins. By default, all
plugin properties will be verified against the previous build step
state. This can be overridden by each plugin if needed.

The implementation is also designed to work with strict plugins
where plugin properties changes may require a repull. In this case,
each plugin must return the list of properties relevant to the pull
step according to the offline build mode.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>