plugins: implement strict/offline plugins (#361)
Instead of having a StrictPlugin subclass, have a 'supports_strict_mode'
classvar that declares whether the Plugin subclass is capable of
operating in 'strict mode'. Implementations should then inspect the
runtime state (via the part info) to see if the methods are being
in regular or strict mode.

This also adds a new method to the superclass, get_pull_commands(). The
default implementation is an empty list to preserve backwards-compat.