packages: don't write info logs in overlays
This commit is a workaround for a current bad interaction between the
way we perform overlay operations (a chroot in a multiprocess-created
subprocess) and the way craft applications indicate progress on the
cli (a spinner in a thread).

The consequence of the bad interaction is that trying to write an info
message to the terminal ends up deadlocking the main thread in the
subprocess because the thread that could unblock it does not exist.

Instead, log the message as 'debug' instead of 'info', and add a main-
process-side 'info' message to communicate that overlay packages are
being fetched.