docs: Link filesets specification from part properties
Was looking for how the `prime` and `stage` part properties
take glob-like inclusion and exclusion patterns, but the
docs didn't point to the fact that these are "filesets".

Add a link to the fileset specification explanation, so that
people looking for all the possible fileset-specific ways of
specifying files to include and exclude can find it easily.

Also, "at least item" -> "at least 1 item"