feat(docs)!: rework bundling of shared docs (#688)
This commit re-organizes the docs/ folder to share *most* of the rst files.
Moving forward, the expectation is that new content will be written inside of
docs/common/craft-parts/ if there is any expectation at all that downstream
applications will want to include it in their own docs. The idea is to keep
to the /common/craft-parts/ folder so that both Craft Parts and applications
can refer to the docs using the same path,  without overlaps/conflicts.

Not everything is shared - some docs are kept in Craft Parts as they are not
meant to be consumed by users of downstream applications. Some examples:

- Class references to Craft-parts internals;
- Docs about using ``craft-parts`` directly from the command line;
- "Meta documentation" about documenting the library itself;
- Front-page and tables of contents that structure Craft Parts' own docs.