fix: use partitions for stage-packages tracking
This localized fix addresses the issue of stage-package tracking failing in
the presence of partitions by looking for the primed files in each partition.

This solves the most common cases but doesn't fix the issue where a given file
might be present in multiple partitions, as this needs a deeper refactoring of
the state tracking (see discussion in #804).

Fixes #804