fix: more closely replicate the subprocess module for backwards compatibility
Most tests and other modules expect `subprocess.run()` to be called, so the kwargs passed into it are given with the expectation that they'll be handled that way. Instead of changing each call one-by-one, this commit changes `fork_utils.run()` to default to the system's stdout/stderr when given `None` instead of not redirecting output at all.