docs: convert interlinks in plugin pages to external refs
Fixes the library side of CRAFT-3890.

- Makes the hosted Craft Parts docs an external source for this doc set.
- All interlinks in the plugin pages, which are often used by apps as
  modular docs, now use intersphinx and treats them as external links.
- Any potentially standalone pages should prefix references to other pages
  with `:external+craft-parts:ref:`. References to labels within the page can
  use simple `:ref:` like normal.

There's a flaw with this approach, though. Only already published pages can
be linked to externally. If for example we add a new plugin page and want to
link pages to it as part of a code change, those links will break beause the
new page hasn't been published yet.