feat(executor): allow capturing error message from scriptlets (#966)
* change craftctl channel to socket

Change to craftctl communication channel from named pipes to unix
sockets to make it easier to consolidate the ctl server and error
stream capture.

* allow using process selector with custom handlers

Process execution runs a selector to obtain data from i/o streams
to obtain the isolated error output. Allow this selection loop run
on a caller-specified selector with handlers registered to also
process interactions with craftctl.

* change scriptlet environment script name

Step execution environment is step-specific, and the existing
environment.sh script is for the build step only. Rename the file
containing the scriptlet environment to avoid confusion with the
existing file (which may potentially have a different payload).

* force select selector

Using DefaultSelector fails in jammy with Python 3.10. It seems that
this specific version picks a different selector, or there's some
problem with SelectSelector that prevents it from working as expected.

---------

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>