feat: 'self-contained' builds for Maven plugins
This commit introduces the "self-contained" build-attribute. This introduction
only applies to Maven-based ("maven" and "maven-use") plugins for now, and when
enabled causes the plugins to *always* rewrite the required version of Maven
sources based on those artifacts available in the backstage (produced by
maven-use parts) and in /usr/share/maven-repo (produced by Maven-related archive
packages).

A side effect of this is that Maven builds now *always* discovers artifacts in
the backstage. This is desirable as a default behavior anyway, as the expectation
is that parts will consume from the shared "stage" (and now also the backstage)
when building. Note that if the build isn't self-contained, the discovery of the
artifacts in the backstage doesn't mean that they said artifacts *will* be used,
just that they *can* be used.