fix: Prime content from overlay when split into partitions (#1123)
Rework migrating content from the overlay at the PRIME step.
Due to the new feature to distribute content from the overlay to partitions, the content of the default stage directory may not fully reflect the content from the overlay. Consequently, the priming cannot rely on the overlay anymore.

Adopt a new approach relying on the content of the stage directory and the stage migration state (to get the list of files/directories coming from the overlay).
This approach is also closer to the standard way to migrate content.

Also make sure to not overwrite existing overlay migration prime state files.

---------

Signed-off-by: Paul Mars <paul.mars@canonical.com>