feat: Use the first partition in the list as the default (#1144)
Enable setting a custom name as the first partition in the partition list provided to the lifecycle, and consider it the default one.
Only the first partition in the list can be named 'default'.
Crafters can still reference the default partition with (default) in the project file.
Migration states now also track the partition name to properly track migrated content per partition.

---------

Signed-off-by: Paul Mars <paul.mars@canonical.com>