feat(maven-use): consume binaries directly
This commit give the `maven-use` plugin the ability to consume, as "source",
artifacts that were previously built by Maven. This ability is constrained to
the existence of a "maven-use" top-level directory in the part's source, which
indicates that the plugin should use those binaries directly instead of
attempting to build from Java source.

To illustrate this behavior, the new integration test builds two "projects":
the first, "binaries-source", builds two Java dependencies using the maven-use
plugin and copies the generated artifacts as "sources" for the second project,
"binaries-consume". This project's third and final part then uses regular Java
sources with the Maven plugin to create a fat "application" jar.