fix(overlay): detect conflicts between overlay and stage
This commit enhances the collision detection in the Stage step to also take
into account contents coming from the overlay. The standard collision detection
was done by comparing the contents of the install directories for the existing
parts - this behavior persists, but now we *also* take into account the contents
of each overlay-enabled part's layer directory, taking into account the layer
stack to only report collisions between items that would *actually* be staged.