feat: dynamically detect the package repository type
The package repository type was historically detected once at the start
of execution, based on existing system files. A new feature in Craft
Parts will allow creation of the base system in a part, and the base
system must now be detected dynamically (because at start of execution
the base system may be potentially empty). Additionally, the repository
type must be re-detected inside a chroot since it may contain a
different system compared to the host where Craft Parts is running.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>