feat(plugins): new usrmerge-ready Python plugin
The original Python plugin is incompatible with usrmerged systems because the
virtual environment directory structure that it needs is very hardcoded. This
new plugin does not use virtual environments; instead, during the build the
plugin uses the part's install dir as the base for "pip install --user" calls.

A sitecustomize file is provided so that at runtime the interpreter can pick
up the packages installed by pip. Since this sitecustomize has to be discoverable,
this means that the plugin assumes that the interpreter will be included in the
primed payload (which is true for Rockcraft, and will be true for Snapcraft
starting with core26).

This new plugin is not registered by default; applications wanting to use it,
typically in conjunction with usrmerged install dirs, need to explicitly
register it *over* the default Python plugin.