feat: allow creating usrmerged dirs by default (#1262)
This commit introduces a "usrmerged_by_default" parameter to the LifecycleManager
so that applications can opt into a new behavior where a part's install dir gets
an usrmerged directory structure by default. This lets the lifecycle create a set
of files that is better suited for modern Ubuntu bases, but is a breaking change
for some plugins and specially for user-provided "override-build" scripts.

Two build-attributes are also introduced to let users have finer, per-part control:

- "enable-usrmerge" can be used to have the directory structure created for the
  part even if "usrmerged_by_default" is False;
- "disable-usrmerge" can be used to disable the structure creation even if
  "usrmerged_by_default" is True.

Note that the behavior is disabled for the `dump` and `nil` plugins as a special case; it can
still be opted-into via "enable-usrmerge" but the new usmerge structure would be too
surprising (and possibly breaking) for very common cases like:

```yaml
config-files:
  plugin: dump
  source: config/
  organize:
    "*": etc/config/
```