feat: make cache layer level configurable (#1263)
In the overlay stack, place the package cache layer below the
first level that does not organize to the overlay after a part
organizes to the overlay. If no parts organize to the overlay
the cache layer is placed immediately above the base layer,
retaining compatibility with the behavior before implementing
organization to the overlay.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>