fix: ignored outdated files should still be pulled
During the execution of the lifecycle there are files or directories
that should not be pulled, such as the `parts` or `stage`. There are
also files that should be pulled and ignored by the outdated file
checker, such as `spread.yaml`. If we don't pull the latter projects
that set `version` using `git describe --dirty` will always describe
the state of the repository as dirty (because the files are part of
the repository but are missing because they were not pulled).

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>